refactor(sdk): decouple defi llama api call
