+++ packages/sdk/src/functions/index.ts
@@ -4,4 +4,5 @@ export { findPermitAction } from './findPermitAction';
 export { getPermitDigest } from './getPermitDigest';
 export { getPreviewActions } from './getPreviewActions';
 export { getPreviewRoutingDetails } from './getPreviewRoutingDetails';
+export { llamaFinancials } from './llamaFinancials';
 export { waitForTransaction } from './waitForTransaction';
